fix merge, cleanup
