--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TurnoTesto.java
@@ -1,16 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-import org.firstinspires.ftc.teamcode.auto.ChassisConfig;
-import org.firstinspires.ftc.teamcode.autoOp.LongBeanis1;
-import org.firstinspires.ftc.teamcode.autoOp.TurnTest;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-//Beanis = Square
-//Hagrids = Slash
-
-@Autonomous(name="Turn Test Quarter", group="1Testing")
-public class TurnoTesto extends TurnTest {
-    public TurnoTesto() {
-        super(ChassisConfig.forTileRunner());
-    }
-}
