--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TylerClawTest.java
@@ -1,14 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-import org.firstinspires.ftc.teamcode.auto.ChassisConfig;
-import org.firstinspires.ftc.teamcode.autoOp.ClawTest;
-import org.firstinspires.ftc.teamcode.autoOp.OniChan;
-
-@Autonomous(name="ClawTest", group="ZZTesting")
-public class TylerClawTest extends ClawTest {
-    public TylerClawTest() {
-        super(ChassisConfig.forTileRunner());
-    }
-}
