+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoOpTest/ClawTest.java
@@ -1,4 +1,6 @@
-package org.firstinspires.ftc.teamcode.autoOp;
+package org.firstinspires.ftc.teamcode.autoOpTest;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.firstinspires.ftc.teamcode.auto.ChassisConfig;
 import org.firstinspires.ftc.teamcode.auto.ChassisStandard;
@@ -6,12 +8,8 @@ import org.firstinspires.ftc.teamcode.auto.ChassisStandard;
 /**
  *
  */
-public abstract class ClawTest extends ChassisStandard {
-
-
-    public ClawTest(ChassisConfig config) {
-        super(config);
-    }
+@Autonomous(name="Claw Test", group="ZZTesting")
+public class ClawTest extends ChassisStandard {
 
     /**
      * Code to run REPEATEDLY after the driver hits PLAY but before they hit STOP
