+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoOpTest/SafetyPatrol.java
@@ -1,16 +1,15 @@
-package org.firstinspires.ftc.teamcode.autoOp;
+package org.firstinspires.ftc.teamcode.autoOpTest;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.firstinspires.ftc.teamcode.auto.ChassisStandard;
 import org.firstinspires.ftc.teamcode.auto.ChassisConfig;
 
 /**
- * This just runs from the position closest to the crater, into the crater.
+ *
  */
-public abstract class SafetyPatrol extends ChassisStandard {
-
-    public SafetyPatrol(ChassisConfig config) {
-        super(config);
-    }
+@Autonomous(name="SafetyPatrol Test", group="ZZTesting")
+public class SafetyPatrol extends ChassisStandard {
 
     /**
      * Code to run REPEATEDLY after the driver hits PLAY but before they hit STOP
