+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoOpTest/ShortTrayTurnTest.java
@@ -1,22 +1,20 @@
-package org.firstinspires.ftc.teamcode.autoOp;
+package org.firstinspires.ftc.teamcode.autoOpTest;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.firstinspires.ftc.teamcode.auto.ChassisConfig;
 import org.firstinspires.ftc.teamcode.auto.ChassisStandard;
 
 /**
- * This just runs from the position closest to the crater, into the crater.
+ *
  */
-public abstract class ShortTrayTurnTest extends ChassisStandard {
+@Autonomous(name="Short Tray Turn Test", group="ZZTesting")
+public class ShortTrayTurnTest extends ChassisStandard {
 
-
-    public ShortTrayTurnTest(ChassisConfig config) {
-        super(config);
-    }
 
     /**
      * Code to run REPEATEDLY after the driver hits PLAY but before they hit STOP
      */
-
     public void loop () {
 
         if (madeTheRun == false) {
@@ -29,6 +27,8 @@ public abstract class ShortTrayTurnTest extends ChassisStandard {
 
             madeTheRun = true;
         }
+
+        printStatus();
     }
 }
